# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "API for chess tournament administration system.",
        "contact": {
            "email": "xdudasko@stuba.sk, xdemeterp@stuba.sk"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://mtaa_chess_api.test/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/matches": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get list of matches",
                "description": "Returns list of matches",
                "operationId": "getMatchesList",
                "parameters": [
                    {
                        "name": "round",
                        "in": "query",
                        "description": "Round number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "table": 1,
                                        "white": 3,
                                        "black": 8,
                                        "result": null,
                                        "round": 5
                                    },
                                    {
                                        "table": 2,
                                        "white": 4,
                                        "black": 9,
                                        "result": 1,
                                        "round": 5
                                    },
                                    {
                                        "table": 3,
                                        "white": 5,
                                        "black": 10,
                                        "result": 2,
                                        "round": 5
                                    },
                                    {
                                        "table": 4,
                                        "white": 6,
                                        "black": 11,
                                        "result": 0,
                                        "round": 5
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Matches"
                ],
                "summary": "Create new round",
                "description": "Returns new round data",
                "operationId": "storeMatch",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "table": 1,
                                        "white": 3,
                                        "black": 8,
                                        "result": null,
                                        "round": 5
                                    },
                                    {
                                        "table": 2,
                                        "white": 4,
                                        "black": 9,
                                        "result": 1,
                                        "round": 5
                                    },
                                    {
                                        "table": 3,
                                        "white": 5,
                                        "black": 10,
                                        "result": 2,
                                        "round": 5
                                    },
                                    {
                                        "table": 4,
                                        "white": 6,
                                        "black": 11,
                                        "result": 0,
                                        "round": 5
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/matches/{id}": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get list of player's matches. (User with id 2 in the example.)",
                "description": "Returns list of player's matches",
                "operationId": "getPlayerMatches",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Player id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "table": 1,
                                        "white": 2,
                                        "black": 8,
                                        "result": 3,
                                        "round": 1
                                    },
                                    {
                                        "table": 2,
                                        "white": 9,
                                        "black": 2,
                                        "result": 1,
                                        "round": 2
                                    },
                                    {
                                        "table": 3,
                                        "white": 2,
                                        "black": 10,
                                        "result": 2,
                                        "round": 3
                                    },
                                    {
                                        "table": 4,
                                        "white": 11,
                                        "black": 2,
                                        "result": 0,
                                        "round": 4
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Matches"
                ],
                "summary": "Update existing match",
                "description": "Returns updated match data",
                "operationId": "updateMatch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Match id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "result",
                        "in": "query",
                        "description": "Match result",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Invalid request body.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, invalid request body."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}